Cleaned things up
Cleaned things up, reordered the code, added and changed comments, and changed Lexer initialization.
