+++ src/Maho/Text/SourceText.cs
@@ -8,35 +8,27 @@ internal sealed class SourceText
     /// <summary> The source text. </summary>
     private readonly string text;
 
+    /// <summary> The length of the source text. </summary>
+    public int Length => text.Length;
+    /// <summary> The lines of the source text. </summary>
+    public TextLine[] Lines { get; }
+
     /// <summary> Initializes a new instance of the <see cref="SourceText"/> class. </summary>
-    /// <param name="text"> The source text. </param>
+    /// <param name="text"> Source text. </param>
     public SourceText(string text)
     {
         this.text = text;
-        Lines = ParseLines(this, text);
+        Lines = ParseLines();
     }
-
-    /// <summary> Gets the length of the source text. </summary>
-    public int Length => text.Length;
-
-    /// <summary> Gets the lines of the source text. </summary>
-    public TextLine[] Lines { get; }
 
     /// <summary> Gets the character at the specified position. </summary>
     /// <param name="position"> The position of the character. </param>
     /// <returns> The character at the specified position. </returns>
     public char this[int position] => text[position];
 
-    /// <summary> Gets the text within the specified span. </summary>
-    /// <param name="span"> The span of text to get. </param>
-    /// <returns> The text within the specified span. </returns>
-    public string ToString(TextSpan span) => text.Substring(span.Start, span.Length);
-
     /// <summary> Parses the lines of the source text. </summary>
-    /// <param name="sourceText"> The source text. </param>
-    /// <param name="text"> The text to parse. </param>
     /// <returns> An array of <see cref="TextLine"/>. </returns>
-    private static TextLine[] ParseLines(SourceText sourceText, string text)
+    private TextLine[] ParseLines()
     {
         List<TextLine> result = [];
 
@@ -45,55 +37,49 @@ internal sealed class SourceText
 
         while (position < text.Length)
         {
-            var lineBreakWidth = GetLineBreakWidth(text, position);
+            var lineBreakWidth = GetLineBreakWidth(position);
 
             if (lineBreakWidth == 0)
-            {
                 position++;
-            }
             else
             {
-                AddLine(result, sourceText, position, lineStart, lineBreakWidth);
+                AddLine(result, position, lineStart, lineBreakWidth);
                 position += lineBreakWidth;
                 lineStart = position;
             }
         }
 
         if (position >= lineStart)
-        {
-            AddLine(result, sourceText, position, lineStart, 0);
-        }
+            AddLine(result, position, lineStart, 0);
 
         return [.. result];
     }
 
     /// <summary> Adds a line to the result. </summary>
     /// <param name="result"> The result builder. </param>
-    /// <param name="sourceText"> The source text. </param>
     /// <param name="position"> The current position. </param>
     /// <param name="lineStart"> The start of the line. </param>
     /// <param name="lineBreakWidth"> The width of the line break. </param>
-    private static void AddLine(List<TextLine> result, SourceText sourceText, int position, int lineStart, int lineBreakWidth)
+    private void AddLine(List<TextLine> result, int position, int lineStart, int lineBreakWidth)
     {
         var lineLength = position - lineStart;
         var lineLengthIncludingLineBreak = lineLength + lineBreakWidth;
-        var line = new TextLine(sourceText, lineStart, lineLength, lineLengthIncludingLineBreak);
+        var line = new TextLine(this, lineStart, lineLength, lineLengthIncludingLineBreak);
         result.Add(line);
     }
 
     /// <summary> Gets the width of the line break at the specified position. </summary>
-    /// <param name="text"> The text. </param>
     /// <param name="position"> The position. </param>
     /// <returns> The width of the line break. </returns>
-    private static int GetLineBreakWidth(string text, int position)
+    private int GetLineBreakWidth(int position)
     {
-        var c = text[position];
-        var l = position + 1 >= text.Length ? '\0' : text[position + 1];
+        var ch = text[position];
+        var next = position + 1 >= text.Length ? '\0' : text[position + 1];
 
-        if (c == '\r' && l == '\n')
+        if (ch == '\r' && next == '\n')
             return 2;
 
-        if (c == '\r' || c == '\n')
+        if (ch == '\r' || next == '\n')
             return 1;
 
         return 0;
@@ -121,4 +107,13 @@ internal sealed class SourceText
 
         return lower - 1;
     }
+
+    /// <summary> Gets the text as string. </summary>
+    /// <returns> The text. </returns>
+    public override string ToString() => text;
+
+    /// <summary> Gets the text within the specified span. </summary>
+    /// <param name="span"> The span of text to get. </param>
+    /// <returns> The text within the specified span. </returns>
+    public string ToString(TextSpan span) => text.Substring(span.Start, span.Length);
 }
