+++ src/Maho/Text/TextSpan.cs
@@ -1,34 +1,25 @@
 namespace Maho.Text;
 
 /// <summary> Represents a span of text in the source code. </summary>
-/// <param name="start"> The start position of the span. </param>
-/// <param name="length"> The length of the span. </param>
+/// <param name="start"> Start position of the span. </param>
+/// <param name="length"> Length of the span. </param>
 internal readonly struct TextSpan(int start, int length)
 {
-
-    /// <summary> Gets the start position of the span. </summary>
+    /// <summary> The start position of the span. </summary>
     public int Start { get; } = start;
-
-    /// <summary> Gets the length of the span. </summary>
+    /// <summary> The length of the span. </summary>
     public int Length { get; } = length;
-
-    /// <summary> Gets the end position of the span. </summary>
+    /// <summary> The end position of the span. </summary>
     public int End => Start + Length;
 
     /// <summary> Creates a new <see cref="TextSpan"/> from start and end positions. </summary>
     /// <param name="start"> The start position of the span. </param>
     /// <param name="end"> The end position of the span. </param>
     /// <returns> A new <see cref="TextSpan"/>. </returns>
-    public static TextSpan FromBounds(int start, int end)
-    {
-        return new TextSpan(start, end - start);
-    }
+    public static TextSpan FromBounds(int start, int end) => new(start, end - start);
 
     /// <summary> Gets the line number of the start position. </summary>
-    public int GetStartLine(SourceText sourceText)
-    {
-        return sourceText.GetLineIndex(Start);
-    }
+    public int GetStartLine(SourceText sourceText) => sourceText.GetLineIndex(Start);
 
     /// <summary> Gets the column number of the start position. </summary>
     public int GetStartColumn(SourceText sourceText)
@@ -39,10 +30,7 @@ internal readonly struct TextSpan(int start, int length)
     }
 
     /// <summary> Gets the line number of the end position. </summary>
-    public int GetEndLine(SourceText sourceText)
-    {
-        return sourceText.GetLineIndex(End);
-    }
+    public int GetEndLine(SourceText sourceText) => sourceText.GetLineIndex(End);
 
     /// <summary> Gets the column number of the end position. </summary>
     public int GetEndColumn(SourceText sourceText)
